apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Music Agent Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langchain>=0.1.0,<0.2.0",
    "langchain-openai>=0.0.5,<0.1.0",
    "spotipy>=2.23.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "asyncpg>=0.31.0",
    "sqlalchemy>=2.0.45",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
